int statt double für vormerkmengen, Animation "kommt bald wieder" stecken

+++ app/src/main/java/wien/mila/nachschlichten/data/local/dao/PendingItemDao.kt
@@ -12,7 +12,7 @@ data class PendingItemWithArticle(
     val articleName: String,
     val articleEan: String,
     val shelfId: String,
-    val quantity: Double?,
+    val quantity: Int?,
     val createdAt: Long,
     val isDone: Boolean
 )
